Fix unmarshaling into nil fields with no corresponding tag.
Prior to this change, a field of pointer type would be pre-filled with a
non-nil pointer to a zero value. This works, but results in the output having a
spurious pointer.

Instead of pre-allocating, update the unmarshaler to check for unmarshaling
into a value of type **T. Only allocate a value at the point where data are
available to decode into that field.

Ancillary fixes:

- Remove allocation code from the wantPointer=true path.
- Fix an error path that was reporting nil.